refactor: revamp profile and settings screens
