--- app/(tabs)/profile.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import ProfileScreen from '../profile';
-
-export default function ProfileTab() {
-  return <ProfileScreen />;
-}
