+++ app/(tabs)/profile/_layout.tsx
@@ -1,0 +1,6 @@
+import { Stack } from 'expo-router';
+
+export default function ProfileLayout() {
+  return <Stack screenOptions={{ headerShown: false }} />;
+}
+
