--- app/(tabs)/settings.tsx
@@ -1,23 +1,0 @@
-import React from 'react';
-import { View, Text, StyleSheet } from 'react-native';
-
-export default function SettingsScreen() {
-  return (
-    <View style={[styles.container, { backgroundColor: '#101828' }]}>  {/* Match login screen color */}
-      <Text style={styles.text}>الإعدادات</Text>
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#0f172a',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  text: {
-    color: '#fff',
-    fontSize: 18,
-  },
-});
